backend/requirements.txt
Flask==2.2.5
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.6
pandas==2.0.3
openpyxl==3.1.2
werkzeug==2.2.3
numpy==1.24.3